pyproject.toml
[project]
name = "anagnorisis"
version = "0.1.0"
description = "An AI-powered D&D-style game with real-time events and multiplayer support"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Narek", email = "your-email@example.com" }
]

dependencies = [
    # Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    
    # AI
    "anthropic>=0.18.0",
    
    # Client TUI
    "textual>=0.47.0",
    "rich>=13.7.0",
    
    # Shared utilities
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
# These let you run `anagnorisis-server` and `anagnorisis-client` from terminal
anagnorisis-server = "server.main:main"
anagnorisis-client = "client.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "client*", "shared*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
